refactor : calculate and store kebutuhan pegawai
this commit is enables the system to calculate and store the value of 'total_waktu_penyelesaian_tugas'  and 'kebutuhan_pegawai'  in 'abk_jabatan' table when a user stores a detail abk. this commit also replace all the code in which 'total_waktu_penyelesaian_tugas' and 'kebutuhan_pegawai' is displayed to users.

changes
- add 'total_waktu_penyelesaian_tugas' column to AbkJabatan migration
- update AbkController@storeDetailAbk to calculate and store 'total_waktu_penyelesaian_tugas'  and 'kebutuhan_pegawai'
- update AbkController@showJabatan and AbkController@editJabatan to use 'total_waktu_penyelesaian_tugas'  and 'kebutuhan_pegawai' from abk jabatan table
- update /abk/jabatan/edit.blade.php and /abk/jabatan/show.blade.php to use passed wpt and kebutuhan_pegawai data
- include abk_jabatan data to detail.abk.store route in web.php